Time: 53 ms (55.53%), Space: 18.8 MB (98.14%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -1,16 +1,16 @@
 class Solution {
 public:
     int minEatingSpeed(vector<int>& piles, int h) {
-        long long start = 1;
-        long long end = *max_element(piles.begin(), piles.end());
+        int start = 1;
+        int end = *max_element(piles.begin(), piles.end());
         
         // if(piles.size() == 1){
         //     return piles.front();
         // }
         
         while(start <= end){
-            long long mid = start + (end - start)/2;
-            long long time = 0;
+            int mid = start + (end - start)/2;
+            int time = 0;
             for(auto bananas: piles){
                 time += ((bananas/mid)+ ((bananas%mid) != 0));
                 if(time > h){
